Fixes and Dev
Changed method in Lane so vehicles are added to the back of the lane, not the front/ right behind another vehicle.
Junction now creates lanes that share the same traffic light across each road.
Vehicle movement should be completed but has not been tested.
Fixed missing 'void' in TrafficLightConfig class.