Pre-integration Work
Added a triple class to store a 3 length tuple of data.
Added a way to get vehicles that have moved from their entry lanes, so they can be animated.
Changed vehicles to take in a string direction (overloaded method so that it is not necessary), to support the above method.
Added more comments.
Tweaked simulation tests.